Memory leak and compiler warnings fixes:
- memory leaks in pqi/cleanupxpgp.cc , 2 new char[] => 2 delete[]
- reordering some constructors' initializing order
- char* myString = "foo" constructs must be prefixed by "const"
- using size_t for variable types when dealing with STL


git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@3079 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/dbase/rsexpr.cc
@@ -82,7 +82,7 @@ bool PathExpression::eval(FileEntry *file){
 bool ExtExpression::eval(FileEntry *file){
 	std::string ext;
 	/*Get the part of the string after the last instance of . in the filename */
-	unsigned int index = file->name.find_last_of('.');
+	size_t index = file->name.find_last_of('.');
 	if (index != std::string::npos) {
 		ext = file->name.substr(index+1);
 		if (ext != "" ){
